pyproject.toml
[project]
name = "project-boilerplate"
version = "0.1.0"
description = ""
authors = [
    {name = "H2Shami",email = "H2Shami@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pypdf2 (>=3.0.1,<4.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "openai (>=1.106.1,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "anthropic (>=0.66.0,<0.67.0)",
    "pytesseract (>=0.3.13,<0.4.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
